Extracted to a JITM to a custom view
